CA-138228: Reduce memory usage not using frame buffer copies if there are no clients
Reduce memory usage for VNC.
VNC require the allocation of the entire framebuffer in order to send
the differences instead of the full image.
However this memory is allocated always even if there is no VNC
connection. This patch avoid such allocation if not needed.
Memory is allocated on demand if requested.
The surface is marked with a specific flag (QEMU_LAZY_FLAG) to tell that
the default allocator is used.

If the console is selected cirrus driver update the VGA memory. This cause
video memory to be allocated even if nobody is interested to.

Deselecting it if we don't have any VNC connection cause to not allocate
and update this memory.