CA-142061: Fix possible race condition executing command
Command from xenstore are written within a transaction writing command
and parameter. This assure Qemu cannot read the parameter of old command.
However to acknowledge we got a command we remove the command from xenstore.
This can be create a race if we remove the command before reading the
parameter as after the command removal toolstack could set again command
and parameter before we read the parameter for the command we got.