CA-145892: vga: map VRAM when BAR is first enabled
On Xen, the frames for the VRAM are populated during device
initialization at 0xFF00_0000.  They are later mapped to a different
guest physical address (the linear frame buffer address) set from BAR0
(normally 0xF100_0000).

This mapping is done when the guest enables the display device by
setting the ENABLED bit in the VBE ENABLE register.

Grub, however, attempts to access the linear frame buffer /before/ the
device has been enabled in this way (possbily because of the order it
probes for different display devices).  Grub accesses the frame buffer
> 32,000 times and each time, qemu tries to map frames (trying to
populate the map cache with a batch of 256 frames each time which are
not present in the p2m.  This is expensive and significantly slows
down boot.

The linear frame buffer should be accessible when the BAR is enabled,
so map it when the BAR is first enabled.  This does not handle the BAR
being moved later on (e.g., by the guest OS) but this does not happen
in practice.