CA-145892: vga: don't re-map VRAM on restore
On restore the generic PCI device state is loaded before any of the
VGA-specific state.  In particular, s->vram_gmfn is zero when
pci_device_load() calls vga_map() and set_vram_mapping().  This causes
the beginnning of the frame buffer BAR to point to PFNs 0x00 to 0x9F
(the remap fails at PFN 0xA0 which is the hole for the VGA memory
region).

Writes to the frame buffer will thus corrupt the lower 640 kiB of
guest memory causing the guest to quickly triple fault.

Ensure we do not map the VRAM if we don't know where it is yet (i.e.,
if s->vram_gmfn == 0).  This works because the physmap is actually
preserved during restore and the frame buffer BAR does not need to be
re-mapped.