CA-147808: Prevent clients missing frame updates
It is possible for the following situation to occur:
Client requests update -->
                            <-- Server sends update
                                Display changes
Client requests update -->

Prevent the client missing this frame update by remembering when the
display changes but the client hasn't requested an update and sending
the changes to the client when it next requests an update.

This is particularly noticeable on high latency links or slow clients
(like XenCenter).